+++ packages/react-api/src/index.d.ts
@@ -1,4 +1,5 @@
 export { executeSQL } from './api/SQL';
+export { ldsGeocode } from './api/lds';
 
 export { default as useCartoLayerProps } from './hooks/useCartoLayerProps';